Move useToast hook to hooks directory
