+++ src/hooks/useToast.tsx
@@ -1,5 +1,5 @@
 // Inspired by react-hot-toast library
-import type { ToastActionElement, ToastProps } from '@/components/ui/toast';
+import type { ToastActionElement, ToastProps } from '@/components/ui/toast.tsx';
 import { ReactNode, useEffect, useState } from 'react';
 
 const TOAST_LIMIT = 1;
